+++ tests/testthat/test-l_cols_width.R
@@ -213,3 +213,140 @@ test_that("cols_width() works correctly in LaTeX output tables", {
 
   expect_match(tbl_latex_tabul, "\\\\begin\\{tabular\\*\\}\\{\\\\linewidth\\}\\{@\\{\\\\extracolsep\\{\\\\fill\\}\\}>\\{\\\\raggedright\\\\arraybackslash\\}p\\{\\\\dimexpr 37.50pt -2\\\\tabcolsep-1.5\\\\arrayrulewidth\\}|>\\{\\\\raggedright\\\\arraybackslash\\}p\\{\\\\dimexpr 56.25pt -2\\\\tabcolsep-1.5\\\\arrayrulewidth\\}|>\\{\\\\raggedleft\\\\arraybackslash\\}p\\{\\\\dimexpr 112.50pt -2\\\\tabcolsep-1.5\\\\arrayrulewidth\\}>\\{\\\\raggedleft\\\\arraybackslash\\}p\\{\\\\dimexpr 150.00pt -2\\\\tabcolsep-1.5\\\\arrayrulewidth\\}\\}")
 })
+
+
+test_that("cols_width() works correctly in LaTeX output tables when row_group_as_column is TRUE or FALSE", {
+
+  tbl <-
+    dplyr::tibble(
+      row = 1:6,
+      group = c(rep("Group A<br>test", 3), rep("Group B", 3)),
+      vals = 1:6
+    )
+
+  tbl_rgac <- tbl |>
+    gt(
+      rowname_col = "row",
+      groupname_col = "group",
+      row_group_as_column = TRUE,
+      process_md = TRUE
+    ) |>
+    fmt_markdown() |>
+    cols_width(
+      row_group() ~ px(200),
+      stub() ~ px(100),
+      vals ~ px(50)
+    ) |>
+    tab_style(style = list(cell_text(align = "right")), cells_stub()) |>
+    as_latex() |>
+    as.character()
+
+  expect_length(tbl_rgac, 1)
+
+  ## rowgroup column
+  expect_match(tbl_rgac, "\\\\multirow\\[t\\]\\{3\\}\\{=\\}\\{\\\\shortstack\\[l\\]\\{\\\\parbox\\{\\\\linewidth\\}\\{Group A \\\\\\\\test\\}\\}\\}")
+  expect_match(tbl_rgac, "\\\\multirow\\[t\\]\\{3\\}\\{=\\}\\{Group B\\}")
+
+  ## styling stub
+  expect_match(tbl_rgac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {1}}}", fixed = TRUE)
+
+
+
+  tbl_rgnac <- tbl |>
+    gt(
+      rowname_col = "row",
+      groupname_col = "group",
+      row_group_as_column = FALSE,
+      process_md = TRUE
+    ) |>
+    fmt_roman(columns = stub()) |>
+    fmt_markdown() |>
+    cols_width(
+      row_group() ~ px(200),
+      stub() ~ px(100),
+      vals ~ px(50)
+    ) |>
+    tab_style(style = list(cell_text(align = "right")), cells_stub()) |>
+    as_latex() |>
+    as.character()
+
+  expect_length(tbl_rgnac, 1)
+
+  ## rowgroup column
+  expect_match(tbl_rgnac, "\\\\multicolumn\\{2\\}\\{>\\{\\\\raggedright\\\\arraybackslash\\}m\\{262.5pt\\}\\}\\{\\\\shortstack\\[l\\]\\{\\\\parbox\\{\\\\linewidth\\}\\{Group A \\\\\\\\test\\}\\}\\}")
+  expect_match(tbl_rgnac, "\\\\multicolumn\\{2\\}\\{>\\{\\\\raggedright\\\\arraybackslash\\}m\\{262.5pt\\}\\}\\{\\\\parbox\\{\\\\linewidth\\}\\{Group B\\}\\}")
+
+  ## styling stub
+  expect_match(tbl_rgnac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {1}}}", fixed = TRUE)
+
+})
+
+
+test_that("cols_width() works correctly in LaTeX output tables when row_group_as_column is TRUE or FALSE, multistub", {
+
+  tbl <-
+    dplyr::tibble(
+      row1 = 1:6,
+      row2 = LETTERS[1:6],
+      group = c(rep("Group A<br>test", 3), rep("Group B", 3)),
+      vals = 1:6
+    )
+
+  tbl_rgac <- tbl |>
+    gt(
+      rowname_col = c("row1","row2"),
+      groupname_col = "group",
+      row_group_as_column = TRUE,
+      process_md = TRUE
+    ) |>
+    fmt_markdown() |>
+    cols_width(
+      row_group() ~ px(200),
+      stub() ~ px(100),
+      vals ~ px(50)
+    ) |>
+    tab_style(style = list(cell_text(align = "right")), cells_stub()) |>
+    as_latex() |>
+    as.character()
+
+  expect_length(tbl_rgac, 1)
+
+  ## rowgroup column
+  expect_match(tbl_rgac, "\\\\multirow\\[t\\]\\{3\\}\\{=\\}\\{\\\\shortstack\\[l\\]\\{\\\\parbox\\{\\\\linewidth\\}\\{Group A \\\\\\\\test\\}\\}\\}")
+  expect_match(tbl_rgac, "\\\\multirow\\[t\\]\\{3\\}\\{=\\}\\{Group B\\}")
+
+  ## styling stub
+  expect_match(tbl_rgac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {1}}}", fixed = TRUE)
+  expect_match(tbl_rgac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {A}}}", fixed = TRUE)
+
+
+
+  tbl_rgnac <- tbl |>
+    gt(
+      rowname_col = c("row1","row2"),
+      groupname_col = "group",
+      row_group_as_column = FALSE,
+      process_md = TRUE
+    ) |>
+    fmt_roman(columns = stub()) |>
+    fmt_markdown() |>
+    cols_width(
+      row_group() ~ px(200),
+      stub() ~ px(100),
+      vals ~ px(50)
+    ) |>
+    tab_style(style = list(cell_text(align = "right")), cells_stub()) |>
+    as_latex() |>
+    as.character()
+
+  expect_length(tbl_rgnac, 1)
+
+  ## rowgroup column
+  expect_match(tbl_rgnac, "\\\\multicolumn\\{3\\}\\{>\\{\\\\raggedright\\\\arraybackslash\\}m\\{337.5pt\\}\\}\\{\\\\shortstack\\[l\\]\\{\\\\parbox\\{\\\\linewidth\\}\\{Group A \\\\\\\\test\\}\\}\\}")
+  expect_match(tbl_rgnac, "\\\\multicolumn\\{3\\}\\{>\\{\\\\raggedright\\\\arraybackslash\\}m\\{337.5pt\\}\\}\\{\\\\parbox\\{\\\\linewidth\\}\\{Group B\\}\\}")
+
+  ## styling stub
+  expect_match(tbl_rgnac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {1}}}", fixed = TRUE)
+  expect_match(tbl_rgnac, "\\multicolumn{1}{>{\\raggedleft\\arraybackslash}m{\\dimexpr 75.00pt -2\\tabcolsep-1.5\\arrayrulewidth}}{\\parbox{\\linewidth}{\\raggedleft {A}}}", fixed = TRUE)
+
+})